+++ libretaDigital/src/main/java/com/libretaDigital/interfaces/AssistControlFacade.java
@@ -1,5 +1,6 @@
 package com.libretaDigital.interfaces;
 
+import java.math.BigInteger;
 import java.util.List;
 
 import com.libretaDigital.entities.Group;
@@ -9,6 +10,6 @@ public interface AssistControlFacade {
 
 	List<Student> getStudentsByCode(String groupCode);
 	
-	List<Group> getGroupsByProfessorId(Long professorId);
+	List<Group> getGroupsByProfessorId(BigInteger professorId);
 	
 }
